Reorganize primary sections (#11)
This is a small PR to reorganize some of the primary sections and add a
few sub-sections that I think we will want to have.


![image](https://github.com/user-attachments/assets/7af9442c-bc33-4f73-80ef-1877d524309f)

For context, I wanted to go ahead and do this to make it easier for me
to start populating existing content into these sub-sections and pages.