Install and apply `prettier` for formatting (#12)
We use `prettier` for formatting on all GC JS repos (with standard
rulesets), and I would like to use it here too.

This PR adds it, and applies formatting for a few JS/TS files.

I opted to exclude Markdown files from formatting because that is the
core content of this repository and it is very much WIP, so I don't
think we want to enforce standardization there. At least not yet.

I'm opting not to add any precommit rules for formatting (as we do in GC
Explorer and the other GC JS repos), because it seemed overkill for this
repo. We can just add this standard copy to a future `CONTRIBUTING.md`:

> To apply Prettier autoformatting you can configure your text editor to
format on save [according
> to these docs](https://prettier.io/docs/en/editors.html)
> - For VS Code, set `editor.formatOnSave` to true in your editor
config.

If / when we do any React development, we can add `eslint` for linting
too, but didn't see the need just yet.

---------

Co-authored-by: Nicolás Pace <nicopace@gmail.com>