Add a minimal hosting Guardian Connector page (#55)
Honestly, I don't think we're truly ready to produce some the
information requested in #5...

> - Hosting options and requirements (including local hosting)
> - Cost estimates and planning resources
> - Security and data sovereignty considerations

Additionally, I think that things like the following are better off
being read directly from `gc-deploy` for the kind of user persona that
would actually be self-deploying GC:

> - Cloud infrastructure considerations
> - Installation, maintenance and update procedures
> - Using CapRover

So this PR provides minimally viable copy for the "Hosting" page that
largely points the user to the `gc-deploy` repo.

I'm going to say this closes #5, and that we file more targeted issues
later (e.g. in 2026) to flesh this out more.