Add "Common workflows" section with two pages (download all data, use data in QGIS) (#54)
Closes #3. Closes #18.

I went back and forth on whether these should be "Guides" instead of
pages in "Reference". But grounded in https://diataxis.fr/, I felt that
the content I was creating was still moreso the latter.

I think we can also add pages for these issues in "Common Workflows":

* #34
* #33
* #36

As I was putting this together, it occurred to me that we sometimes
write for different user personas. And maybe one thing we can do is
clarify that explicitly at the top of each page. So as to communicate,
for example, "this page is intended for a community organization's
technical data manager". Could be nice to have a page where we define
each, and then refer back to that page for definitions across the
platform.

Thoughts welcome! (Both on this latter point, and "Common Workflows")